+++ README.md
@@ -9,22 +9,7 @@ This repository contains a dark theme for Jenkins.
 
 ## Usage
 
-To install this theme use the [Simple Theme Plugin](https://plugins.jenkins.io/simple-theme-plugin/).
-
-See the plugin documentation for theme installation guidelines.
-
-_Note: replace 'tag-name' with the latest released version:_
-[![GitHub release](https://img.shields.io/github/release/jenkinsci/dark-theme.svg?label=latest)](https://github.com/jenkinsci/dark-theme/releases/latest)
-
-The CSS URLs are:
-
-```text
-https://cdn.jsdelivr.net/gh/jenkinsci/dark-theme@<tag-name>/theme.css
-https://cdn.jsdelivr.net/gh/jenkinsci/dark-theme@<tag-name>/theme-system.css
-```
-
-* theme.css - forces dark theme for all users
-* theme-system.css - respects users system setting for dark / light mode, in most cases this is the theme you should be using
+To install this theme search for 'Dark theme' in your Jenkins update center
 
 or with the [configuration-as-code](https://github.com/jenkinsci/configuration-as-code-plugin) plugin:
